requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
gunicorn==21.2.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.9
cloudinary==1.41.0
python-dotenv==1.0.1
